+++ src/device-types/HABinarySensor.cpp
@@ -127,7 +127,7 @@ uint16_t HABinarySensor::calculateSerializedLength(
 
     uint16_t size = 0;
     size += DeviceTypeSerializer::calculateBaseJsonDataSize();
-    //size += DeviceTypeSerializer::calculateNameFieldSize(name()); // @todo
+    size += DeviceTypeSerializer::calculateNameFieldSize(getName());
     size += DeviceTypeSerializer::calculateUniqueIdFieldSize(uniqueId());
     size += DeviceTypeSerializer::calculateDeviceFieldSize(serializedDevice);
     size += DeviceTypeSerializer::calculateAvailabilityFieldSize(this);
@@ -182,7 +182,7 @@ bool HABinarySensor::writeSerializedData(const char* serializedDevice) const
         DeviceTypeSerializer::mqttWriteConstCharField(Prefix, _class);
     }
 
-    // DeviceTypeSerializer::mqttWriteNameField(name()); // @todo
+    DeviceTypeSerializer::mqttWriteNameField(getName());
     DeviceTypeSerializer::mqttWriteUniqueIdField(uniqueId());
     DeviceTypeSerializer::mqttWriteAvailabilityField(this);
     DeviceTypeSerializer::mqttWriteDeviceField(serializedDevice);