+++ src/device-types/HASensor.cpp
@@ -125,7 +125,7 @@ uint16_t HASensor::calculateSerializedLength(
 
     uint16_t size = 0;
     size += DeviceTypeSerializer::calculateBaseJsonDataSize();
-    //size += DeviceTypeSerializer::calculateNameFieldSize(name()); // @todo
+    size += DeviceTypeSerializer::calculateNameFieldSize(getName());
     size += DeviceTypeSerializer::calculateUniqueIdFieldSize(uniqueId());
     size += DeviceTypeSerializer::calculateDeviceFieldSize(serializedDevice);
     size += DeviceTypeSerializer::calculateAvailabilityFieldSize(this);
@@ -206,7 +206,7 @@ bool HASensor::writeSerializedData(const char* serializedDevice) const
         );
     }
 
-    // DeviceTypeSerializer::mqttWriteNameField(name()); // @todo
+    DeviceTypeSerializer::mqttWriteNameField(getName());
     DeviceTypeSerializer::mqttWriteUniqueIdField(uniqueId());
     DeviceTypeSerializer::mqttWriteAvailabilityField(this);
     DeviceTypeSerializer::mqttWriteDeviceField(serializedDevice);